list.c Use "long" instead of "int" for command parameters.
Commit 4e5a6cc044c6c82e6c4d842bf056579b6bc47d48 used int instead of long for the
parameters of the LIST command.  This caused compiler warnings and problems on
64 bit machines.  This change fixes that.

+++ src/language/data-io/list.c
@@ -56,9 +56,9 @@ enum numbering
 
 struct lst_cmd
 {
-  int first;
-  int last;
-  int step;
+  long first;
+  long last;
+  long step;
   const struct variable **v_variables;
   size_t n_variables;
   enum numbering numbering;
